(PC-17986) chore(Home): create a headers folder and move header components to clean the home/component folder
