+++ src/features/home/components/index.ts
@@ -1,4 +1,4 @@
-export { HomeHeader } from './HomeHeader'
+export { HomeHeader } from './headers/HomeHeader'
 export { BusinessModule } from './modules/BusinessModule'
 export { ExclusivityModule } from './modules/ExclusivityModule'
 export { OffersModule } from './modules/OffersModule'
